the whole codebase is being reworked to support utf8 and to be cleaner
