+++ src/frontend/frontend.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <frontend/preprocessor/preprocessor.hpp>
+#include <frontend/lexer/lexer.hpp>
+#include <frontend/parser/parser.hpp>
