+++ src/frontend/lexer/lexer.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <frontend/lexer/utils/char.hpp>
+#include <frontend/lexer/utils/common.hpp>
+
+#include <frontend/lexer/structures/position.hpp>
+#include <frontend/lexer/structures/token_types.hpp>
+#include <frontend/lexer/structures/token.hpp>
+
+#include <frontend/lexer/callbacks.hpp>
+
