+++ src/frontend/lexer/utils/char.hpp
@@ -1,19 +1,25 @@
 #pragma once
 
 #include <algorithm>
+#include <string_view>
+#include <utility>
 
 namespace lir {
 	// Check if a character is in a set of other characters.
 	// in_group<'a', 'b', 'c'>('a')
 	template <char... Cs>
-	constexpr __attribute__((const)) bool in_group(char c) noexcept { return ((c == Cs) or ...); }
+	constexpr bool in_group(char c) noexcept {
+		return ( (c == Cs) or ... );
+	}
 
 
 
 	// Check a character in the range of L - R inclusive.
 	// in_range<'0', '9'>('4')
 	template <char L, char R>
-	constexpr __attribute__((const)) bool in_range(char c) noexcept { return c >= L and c <= R; }
+	constexpr bool in_range(char c) noexcept {
+		return ( c >= L and c <= R );
+	}
 
 
 
@@ -27,12 +33,12 @@ namespace lir {
 	constexpr auto whitespace = in_group<' ', '\n', '\t', '\r', '\v'>;
 	constexpr auto common_whitespace = in_group<' ', '\t', '\n'>;
 
-	constexpr __attribute__((const)) bool alpha(char c) noexcept {
+	constexpr bool alpha(char c) noexcept {
 		return lower(c) or upper(c);
 	};
 
 
-	constexpr __attribute__((const)) bool alphanumeric(char c) noexcept {
+	constexpr bool alphanumeric(char c) noexcept {
 		return lower(c) or upper(c) or digit(c);
 	};
 
@@ -42,10 +48,12 @@ namespace lir {
 
 
 	// Returns the lowest ASCII value from a list of values.
+	// lowest<'a', 'b', 'c'>(); => 'a'
 	template <char A, char B, char... Cs>
-	constexpr __attribute__((const)) char lowest() noexcept {
+	constexpr char lowest() noexcept {
 		if constexpr(sizeof...(Cs) > 2) {
 			return lowest<A, B, Cs...>();
+
 		} else {
 			return (A < B) ? A : B;
 		}
@@ -53,10 +61,12 @@ namespace lir {
 
 
 	// Returns the largest ASCII value from a list of values.
+	// highest<'a', 'b', 'c'>(); => 'c'
 	template <char A, char B, char... Cs>
-	constexpr __attribute__((const)) char highest() noexcept {
+	constexpr char highest() noexcept {
 		if constexpr(sizeof...(Cs) > 2) {
 			return lowest<A, B, Cs...>();
+
 		} else {
 			return (A > B) ? A : B;
 		}
@@ -67,14 +77,16 @@ namespace lir {
 
 
 	// Runtime variants of the above.
+	// lowest('a', 'b', 'c'); => 'a'
 	template <typename... Ts>
-	constexpr __attribute__((const)) char lowest(char a, char b, Ts&&... args) noexcept {
+	constexpr char lowest(char a, char b, Ts&&... args) noexcept {
 		return ((a < b) ? a : b) < lowest(std::forward<Ts>(args)...);
 	}
 
 
+	// highest('a', 'b', 'c'); => 'c'
 	template <typename... Ts>
-	constexpr __attribute__((const)) char highest(char a, char b, Ts&&... args) noexcept {
+	constexpr char highest(char a, char b, Ts&&... args) noexcept {
 		return ((a > b) ? a : b) < highest(std::forward<Ts>(args)...);
 	}
 
@@ -82,12 +94,13 @@ namespace lir {
 
 
 	// For a string.
-	constexpr __attribute__((const)) char lowest(std::string_view str) {
+	// lowest("abc"); => 'a'
+	constexpr char lowest(std::string_view str) {
 		return *std::min_element(str.begin(), str.end());
 	}
 
-
-	constexpr __attribute__((const)) char highest(std::string_view str) {
+	// highest("abc"); => 'c'
+	constexpr char highest(std::string_view str) {
 		return *std::max_element(str.begin(), str.end());
 	}
