+++ src/frontend/parser/structures/visitor.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <variant>
+
+// namespace lir {
+
+//     template<typename... Base>
+//     struct Visitor : Base... {
+//         using Base::operator()...;
+//     };
+
+//     template<typename... T>
+//     Visitor(T...) -> Visitor<T...>;
+
+// }
