+++ src/io/io.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <io/file/file.hpp>
+#include <io/structures/file_stack.hpp>
+#include <io/utils/recv_stdin.hpp>
