+++ src/io/utils/recv_stdin.hpp
@@ -1,11 +1,9 @@
 #pragma once
 
-
 #include <string_view>
 #include <string>
 #include <iostream>
 #include <cstdint>
-
 
 namespace lir {
