--- src/lexer/lexer.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <structures/structures.hpp>
-#include <utils/char.hpp>
-#include <lexer/common.hpp>
-#include <lexer/callbacks.hpp>
-
