+++ src/structures/structures.hpp
@@ -1,10 +1,4 @@
 #pragma once
 
-#include <structures/token_types.hpp>
-#include <structures/token.hpp>
-#include <structures/position.hpp>
 #include <structures/view.hpp>
 #include <structures/char_lookup.hpp>
-#include <structures/file_stack.hpp>
-#include <structures/ast.hpp>
-#include <structures/visitor.hpp>
