+++ src/structures/view.hpp
@@ -6,7 +6,7 @@
 #include <iostream>
 #include <utility>
 
-#include <file/file.hpp>
+#include <io/io.hpp>
 
 namespace lir {
