--- src/structures/visitor.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <variant>
-
-namespace lir {
-
-    template<typename... Base>
-    struct Visitor : Base... {
-        using Base::operator()...;
-    };
-
-    template<typename... T>
-    Visitor(T...) -> Visitor<T...>;
-
-}
