--- src/utils/utils.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <utils/exception.hpp>
-#include <utils/recv_stdin.hpp>
-#include <utils/char.hpp>
